Split up types & typechecker; 'make format'
