+++ src/Juvix/Core/HR/Types.hs
@@ -1,7 +1,7 @@
 module Juvix.Core.HR.Types where
 
-import Juvix.Library
 import Juvix.Core.Usage
+import Juvix.Library
 
 data Term primTy primVal binderTy
   = Star Natural
@@ -9,7 +9,6 @@ data Term primTy primVal binderTy
   | Pi Usage (Term primTy primVal binderTy) (Term primTy primVal binderTy)
   | Lam binderTy (Term primTy primVal binderTy)
   | Elim (Elim primTy primVal binderTy)
-
   deriving (Show, Eq, Generic)
 
 data Elim primTy primVal binderTy
@@ -17,5 +16,4 @@ data Elim primTy primVal binderTy
   | Prim primVal
   | App (Term primTy primVal binderTy) (Elim primTy primVal binderTy)
   | Ann Usage (Term primTy primVal binderTy) (Term primTy primVal binderTy)
-
   deriving (Show, Eq, Generic)
