+++ src/Juvix/Core/IR.hs
@@ -1,3 +1,4 @@
 module Juvix.Core.IR where
 
+import Juvix.Core.IR.Typechecker
 import Juvix.Core.IR.Types
