+++ src/Juvix/Library.hs
@@ -21,11 +21,11 @@ where
 import Capability.Error
 import Capability.Reader
 import Capability.State
-import Numeric.Natural
 import Capability.Stream
 import Capability.Writer
 import Data.String (fromString)
 import qualified Data.Text as T
+import Numeric.Natural
 import Protolude hiding
   ( (:.:),
     Constraint,
